Fixed the Decimal Problem
Fixed the problem where getting doubles from the database was kinda impossible. Now all numbers are displayed as doubles within the expense page.

+++ frontend/mobile/accountability_mobile_app/lib/api/expense_crud.dart
@@ -16,7 +16,7 @@ class ExpenseCRUD {
                 'tripId': tripId,
                 'name': name,
                 'description': description,
-                'cost': cost,
+                'cost': cost.toDouble(),
                 'memberIds': memberIds
               }));
       // If there was an error, don't refresh jwt
